Refactor: Use f32 for any layout relate variable (which is more efficiency)
- Updated `reactive.rs` to implement computed and binding for f32.
- Modified `waterui.h` to include new enum for font weights and updated various struct definitions to use f32 instead of f64.
- Adjusted `icon.rs` to change icon size from f64 to f32.
- Changed `progress.rs` to use f32 for progress value.
- Refactored `filter.rs` to replace f64 with f32 in blur, brightness, contrast, saturation, grayscale, hue rotation, and invert structs.
- Updated `view.rs` to accept f32 for width, height, min/max dimensions.
- Adjusted `rich_text.rs` to use f32 for quote marker height.
- Updated documentation in `05-form.md` to reflect the change from f32 to f64 for sliders.

+++ backends/web/src/widgets/general.rs
@@ -77,7 +77,7 @@ pub fn render_loading_progress() -> WebElement {
 }
 
 /// Render an element with padding applied.
-pub fn render_with_padding<F>(padding: f64, render_fn: F) -> WebElement
+pub fn render_with_padding<F>(padding: f32, render_fn: F) -> WebElement
 where
     F: FnOnce() -> WebElement,
 {
